[FEATURE] Made bounds visible when grid dimensions are specified, minor bug fixes

+++ GameOfLife/src/SimulationEditor.h
@@ -17,8 +17,8 @@ namespace gol
 	class SimulationEditor
 	{
 	public:
-		static constexpr int32_t DefaultCellWidth = 20;
-		static constexpr int32_t DefaultCellHeight = 20;
+		static constexpr float DefaultCellWidth = 20.f;
+		static constexpr float DefaultCellHeight = 20.f;
 	public:
 		SimulationEditor(Size2 windowSize, Size2 gridSize);
